+++ src/pages/projectSettings/tabs/NotificationsTab/getFormSchema.tsx
@@ -8,7 +8,6 @@ import { projectTriggers } from "constants/triggers";
 import { projectSubscriptionMethods as subscriptionMethods } from "types/subscription";
 import { GetFormSchema } from "../types";
 import { radioBoxOptions } from "../utils/form";
-import { SubscriptionField } from "./SubscriptionField";
 import { FormState } from "./types";
 
 export const getFormSchema = (
@@ -24,9 +23,7 @@ export const getFormSchema = (
   } = getNotificationSchema(subscriptionMethods);
 
   return {
-    fields: {
-      subscriptionField: SubscriptionField,
-    },
+    fields: {},
     schema: {
       type: "object" as "object",
       properties: {
@@ -58,9 +55,6 @@ export const getFormSchema = (
                   notification: notificationSchema,
                 },
               },
-              subscriberData: {
-                type: "object" as "object",
-              },
             },
           },
         },
@@ -86,10 +80,6 @@ export const getFormSchema = (
           subscriptionData: {
             event: eventUiSchema,
             notification: notificationUiSchema,
-          },
-          subscriberData: {
-            "ui:field": "subscriptionField",
-            "ui:border": "top",
           },
         },
       },
