+++ src/pages/projectSettings/tabs/NotificationsTab/utils.ts
@@ -143,3 +143,15 @@ export const hasInitialError = (
   }
   return false;
 };
+
+export const generateWebhookSecret = () => {
+  let text = "";
+  const possible =
+    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+
+  for (let i = 0; i < 64; i++) {
+    text += possible.charAt(Math.floor(Math.random() * possible.length));
+  }
+
+  return text;
+};
